[SYCL] Return build options for L0
Implement the `CL_PROGRAM_BUILD_OPTIONS` for `piProgramGetBuildInfo`
in the Level Zero plugin.  The previous comment expressed concern that
we cannot identify the build options "for programs that were built
outside and registered".  I assume this was referring to either the
case when a `pi_program` is created via `piProgramCreateWithBinary`
or via `piextProgramCreateWithNativeHandle`.

However, the OpenCL specification does not require this query to return
the build options in this case.  The spec says:

> Return the build, compile or link options specified by the options
> argument in `clBuildProgram`, `clCompileProgram` or `clLinkProgram`,
> whichever was performed last on the specified program object for
> device.
>
> If build status of the specified program for device is
> `CL_BUILD_NONE`, an empty string is returned.

An OpenCL program that is "built outside and registered" would have
status `CL_BUILD_NONE`, so the query should return an empty string.
Since the PI interface closely follows the OpenCL interface, I assume
the PI query should behave the same way.